backend/requirements.txt
setuptools>=75.0.0,<82
fastapi==0.115.0
uvicorn[standard]==0.32.0
redis==5.2.0
prometheus-fastapi-instrumentator==7.0.0
pydantic==2.9.2
opentelemetry-api==1.27.0
opentelemetry-sdk==1.27.0
opentelemetry-instrumentation-fastapi==0.48b0
opentelemetry-instrumentation-redis==0.48b0
opentelemetry-exporter-otlp-proto-http==1.27.0
